Update Globals.java
activate() changed to avoid warning because of deprecation. Fixed a warning with RequestBody.create, by swapping the arguments as suggested by this URL: https://github.com/square/okhttp/blob/parent-4.0.0/okhttp/src/main/java/okhttp3/RequestBody.kt#L167

configLogger() changed to avoid warning. "ComponentBuilder" should be changed to "ComponentBuilder<?>" The question is if it is possible to narrow down to a specific type? But, for now it's more correct syntax than before.